chore: move popups to separate folders
