+++ Client/SubtitlesApp.MAUI/CustomControls/Popups/SubtitlesSettingsPopup.xaml.cs
@@ -1,0 +1,13 @@
+using CommunityToolkit.Maui.Views;
+using SubtitlesApp.ViewModels.Popups;
+
+namespace SubtitlesApp.CustomControls.Popups;
+
+public partial class SubtitlesSettingsPopup : Popup
+{
+    public SubtitlesSettingsPopup(SubtitlesSettingsPopupViewModel vm)
+    {
+        BindingContext = vm;
+        InitializeComponent();
+    }
+}
