+++ Client/SubtitlesApp.MAUI/CustomControls/Popups/TranslationSettingsPopup.xaml.cs
@@ -1,0 +1,13 @@
+using CommunityToolkit.Maui.Views;
+using SubtitlesApp.ViewModels.Popups;
+
+namespace SubtitlesApp.CustomControls.Popups;
+
+public partial class TranslationSettingsPopup : Popup
+{
+    public TranslationSettingsPopup(TranslationSettingsPopupViewModel vm)
+    {
+        BindingContext = vm;
+        InitializeComponent();
+    }
+}
