--- Client/SubtitlesApp.MAUI/CustomControls/SubtitlesSettingsPopup.xaml.cs
@@ -1,13 +1,0 @@
-using CommunityToolkit.Maui.Views;
-using SubtitlesApp.ViewModels;
-
-namespace SubtitlesApp.CustomControls;
-
-public partial class SubtitlesSettingsPopup : Popup
-{
-	public SubtitlesSettingsPopup(SubtitlesSettingsPopupViewModel vm)
-	{
-		InitializeComponent();
-		BindingContext = vm;
-	}
-}
