--- Client/SubtitlesApp.MAUI/CustomControls/TranslationSettingsPopup.xaml.cs
@@ -1,13 +1,0 @@
-using CommunityToolkit.Maui.Views;
-using SubtitlesApp.ViewModels;
-
-namespace SubtitlesApp.CustomControls;
-
-public partial class TranslationSettingsPopup : Popup
-{
-	public TranslationSettingsPopup(TranslationSettingsPopupViewModel vm)
-	{
-		InitializeComponent();
-		BindingContext = vm;
-	}
-}
