+++ Client/SubtitlesApp.MAUI/ViewModels/PlayerWithSubtitlesViewModel.cs
@@ -15,6 +15,7 @@ using SubtitlesApp.Core.Result;
 using SubtitlesApp.Core.Services;
 using SubtitlesApp.Core.Utils;
 using SubtitlesApp.Interfaces;
+using SubtitlesApp.ViewModels.Popups;
 
 namespace SubtitlesApp.ViewModels;
