+++ Client/SubtitlesApp.MAUI/ViewModels/Popups/LoadingPopupViewModel.cs
@@ -1,5 +1,5 @@
 ﻿using CommunityToolkit.Mvvm.ComponentModel;
 
-namespace SubtitlesApp.ViewModels;
+namespace SubtitlesApp.ViewModels.Popups;
 
 public partial class LoadingPopupViewModel : ObservableObject { }
